+++ src/commonMain/kotlin/at/posselt/pfrpg2e/data/kingdom/settlements/Settlement.kt
@@ -30,7 +30,9 @@ data class Settlement(
     val preventItemLevelPenalty: Boolean,
     val structuresInConstruction: List<Structure>,
     val constructedStructures: List<Structure>,
+    val unpaidStructures: List<Structure>,
     val maximumCivicRdLimit: Int,
+    val settlementActions: Int,
 ) {
     private val totalConsumption = size.consumption - consumptionReduction
     val isOvercrowded = occupiedBlocks > residentialLots