+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [5.5.0] - 2025-06-15
 
+### Added 
+
+* When placing V&K Civic buildings, the structure browser now shows the amount of additional settlement actions that you can take for each settlement
+
 ### Changed
 
 * Changes to V&K 1.1 rules changed how resource dice per settlements are calculated. This required hard coding the RD values; instead of configuring it per settlement type, you now need to toggle a new setting (Settlements Generate Resource Dice) now to enable this behavior.